New queries section 15/01/2025
new queries section in Server - 15/01/2025
