+++ G21_Client_Prototype/bin/.gitignore
@@ -3,4 +3,3 @@
 /controllers/
 /entity/
 /gui/
-/bin/
