+++ G21_Server_Prototype_HW2/bin/.gitignore
@@ -1,4 +1,5 @@
 /Server/
 /common/
 /gui/
-/bin/
+/queries.class
+/queries/
